+++ src/hooks/EditorPauseLayer.cpp
@@ -10,34 +10,6 @@ void HookedEditorPauseLayer::onModify(auto& self) {
     }
 }
 
-#ifndef GEODE_IS_MACOS
-void HookedEditorPauseLayer::onExitEditor(cocos2d::CCObject* sender) {
-    em::log::info("Exited editor, onExitEditor");
-    AudioManager::get().exitEditor();
-    EditorPauseLayer::onExitEditor(sender);
-}
-#else
-void HookedEditorPauseLayer::FLAlert_Clicked(FLAlertLayer* p0, bool btnTwo) {
-    if (p0->getTag() == 1 && btnTwo) {
-        AudioManager::get().exitEditor();
-        em::log::info("Exited editor, FLAlert_Clicked with sender tag 1");
-    }
-    EditorPauseLayer::FLAlert_Clicked(p0, btnTwo);
-}
-#endif
-
-void HookedEditorPauseLayer::onSaveAndExit(cocos2d::CCObject* sender) {
-    em::log::info("Exited editor, onSaveAndExit");
-    AudioManager::get().exitEditor();
-    EditorPauseLayer::onSaveAndExit(sender);
-}
-
-void HookedEditorPauseLayer::onSaveAndPlay(cocos2d::CCObject* sender) {
-    em::log::info("Exited editor, onSaveAndPlay");
-    AudioManager::get().exitEditor();
-    EditorPauseLayer::onSaveAndPlay(sender);
-}
-
 void HookedEditorPauseLayer::onResume(cocos2d::CCObject* sender) {
     // dont allow exiting if the song info popup is on screen
     // in fact this will be when the user presses space, so toggle pause
@@ -45,7 +17,7 @@ void HookedEditorPauseLayer::onResume(cocos2d::CCObject* sender) {
         pop->m_playPauseButton->activate();
         return;
     }
-    
+
     EditorPauseLayer::onResume(sender);
 }
 
@@ -70,11 +42,11 @@ bool HookedEditorPauseLayer::init(LevelEditorLayer* editor) {
         lengthLabel->limitLabelWidth(175.f, 0.5f, .1f);
     }
 
-    
+
     // move label from song stopped at
     // id isnt prefixed with modid but checking both just in case it does in the
     // future
-    
+
     auto node = getChildByID("song-stopped-at");
     if (!node) {
         node = getChildByID("aceinetx.song_stopped_at/song-stopped-at");